Changes to db models, app creation, and view questions
I had to make some changes to the way the app is initialized. Usually you would have to use a __init__.py file to get all of this done but i couldn't undertand how to. So i used something similar. I made another file that holds the db models. In that file, i callled a function from the previously mentioned file. This creates the app and allows the db to use the app. These db models are then imported to every single file since it holds the app instance.