Login Rework_Q&A Addons_HTLM Pages Added
I reworked the login system so that it used flask_login instead of sessions. This let me use some different methods that make life easier. I also changed the db stuff so if you have a past version of the db's used you'll have to delete them to get rid of errors. Besides this I changed the q&a portion to only allow users to post if they are logged in. I also added a create account page where the user can do just that and the info will be added to the users db.