Q&A System Finished
I just made a change that groups questions by class. So only general question will show up on the same page. You can swithc between the classes with the buttons that have the class name you want to go to.

There is a new feature that if a class is empty as in it has no classes. The user will be told as such and a new button will pop up that will allow them to post a new question. This doesn't however post that questoin for that specific class. You still have to chose the class subject when creating the question.