quick fix to features if user isn't logged in
there were some problems with the user not being logged and trying to access some stuff.

The user should be able to see post and comment if they aren't logged in so i change the system to work that way. They still can't post comment or question if they aren't logged in however and can't delete post or comments.

Even if they are redirected to creating a new question or comment, they won't be able to if they aren't logged in.

If they are not logged in and try to comment they are redirected to the login page and told they have to be logged in to comment