Quick Change to Reporting
You can no longer remove a comment/post from it's location on viewing by reporting. Instead it just gets added to a db with reported things and an admin can either deletee the thing or remove the report. Removing the report just removees the thing from the db. Deleeteeing is just getting rid of the thing in all db's