Delete Admin Page, Changes to viewing reported items, change to view other post
I got rid of the admin page since it was ugly and instead added the links to the navbar on top. The dropdown menu for the admins options only shows up if you are an admin.

I also changed the way the reported comments and post are viewed. Not how you view them but they look like the view users page now.

I changed the way the view questions shows up to. There is now some space between each question.