Another set of css changes
Found that the sizing i was initially using was messing up with smaller screens. So i changed that.

I also changed the delete account so that it is more in line with what the login and register pages look like.

I've change the color gradient for the login and register pages. TBH im not a fan of it but it's something to be decided on later.

Finally, i added in the text for the home page since i kind of left it half done last time.