Quick Change to the Q&A Sys.
I got rid of the use one of the routes. I changed it so that when the user wnat to post a question, they have a drop menu of sort show up right above the input box. Makes it so that they don't have to gro through different pages to select a class and post. This also got rid of some butotns.

I've also change some font styling so that font size changes when a page is made smaller. DOens't apply to everything.