Added a page, Some things only show up when logged in or logged out
I added a page to the settings that allows users to view the post that they've made. So go here and signed in as sfransis then you can see all of the post sfransis has made.

Besides this, all of the settings options have been reworked to only show up when logged in/out. So if a user is logged out, settings will show "login" and "create account"

If the user is logged in they will have "logout", "delete account", and "view post".